% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_plot.R
\name{autoplot.spectra}
\alias{autoplot.spectra}
\title{Plot spectra}
\usage{
\method{autoplot}{spectra}(
  object,
  subset = NULL,
  x = .data$energy_kev,
  y = .data$cps,
  group = seq_len(nrow(object)),
  ...,
  facet = NULL,
  nrow = NULL,
  ncol = NULL,
  scales = "free"
)
}
\arguments{
\item{object}{A \link{xrf_spectra}.}

\item{subset}{An expression to subset the unnested data.}

\item{x, y}{Columns in .spectra to be used for plotting}

\item{group}{Override the group aesthetic}

\item{...}{Additional \link[ggplot2]{aes} mappings.}

\item{facet}{An expression to facet by}

\item{nrow, ncol, scales}{Passed to \link[ggplot2]{facet_wrap}.}
}
\value{
A ggplot2 plot.
}
\description{
Plot spectra
}
