% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/units.R
\name{energy}
\alias{energy}
\alias{wavelength}
\title{Wavelenth/Energy X-Ray conversion}
\usage{
energy(wavelength_nm)

wavelength(energy_kev)
}
\arguments{
\item{wavelength_nm}{The wavelength, in nanometers}

\item{energy_kev}{The energy, in keV}
}
\description{
Wavelenth/Energy X-Ray conversion
}
\examples{
# Fe:
energy(0.1936) # from Wiki
wavelength(6.40) # from Kenna et al. 2011

}
