% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{new_spectra}
\alias{new_spectra}
\alias{validate_spectra}
\alias{xrf_despectra}
\title{Create, validate spectra}
\usage{
new_spectra(x, subclass = character(0))

validate_spectra(x)

xrf_despectra(x)
}
\arguments{
\item{x}{An object}

\item{subclass}{Additional classes that should be included}
}
\value{
x, reclassed and/or validated
}
\description{
Create, validate spectra
}
