% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectra.R
\name{read_spec_panalytical}
\alias{read_spec_panalytical}
\title{Read a Panalytical spectra file into a spectra object}
\usage{
read_spec_panalytical(path)
}
\arguments{
\item{path}{A filename or vector of filenames that correspond to Panalytical spectrum files
(usually a .mp2 file)}
}
\value{
A spectra object
}
\description{
Read a Panalytical spectra file into a spectra object
}
\examples{
pan_dir <- system.file("spectra_files/Panalytical", package = "xrf")
read_spec_panalytical(file.path(pan_dir, "Panalytical_2017-08-1632-Omnian.mp2"))

}
