% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{read_xrf_example}
\alias{read_xrf_example}
\title{Read XRF example spectra}
\usage{
read_xrf_example(..., .dir = c("Panalytical"), .which = TRUE)
}
\arguments{
\item{...}{Used to \link[dplyr]{filter} spectra}

\item{.dir}{The subdirectory from which to read}

\item{.which}{Used to subset files before they are read. Use TRUE for all.}
}
\value{
A spectra tibble
}
\description{
Read XRF example spectra
}
\examples{
read_xrf_example(.which = 1:10)
read_xrf_example(SampleIdent == "oreas 22d")

}
