% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_plot.R
\docType{data}
\name{stat_xrf_peaks}
\alias{stat_xrf_peaks}
\alias{StatXrfPeaks}
\title{Add XRF Peaks to a ggplot}
\format{
An object of class \code{StatXrfPeaks} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
stat_xrf_peaks(
  mapping = NULL,
  data = NULL,
  position = "identity",
  ...,
  element_list = "lake_sediment",
  energy_subset = NULL,
  epsilon = 0.01,
  box.padding = 0.5,
  size = 2,
  inherit.aes = TRUE,
  show.legend = FALSE
)

StatXrfPeaks
}
\arguments{
\item{mapping}{Mapping for layer (x and y required)}

\item{data}{Data for layer}

\item{position}{Position for layer}

\item{...}{Passed to \link[ggrepel]{geom_label_repel}. \code{nudge_y} is particularly useful.}

\item{element_list}{An element list (sanitized using \link{xrf_element_list})}

\item{energy_subset}{Used to subset \link{x_ray_xrf_energies}.}

\item{epsilon}{Minimum peak height to label}

\item{box.padding, size}{Passed to \link[ggrepel]{geom_label_repel}}

\item{inherit.aes}{Inherit aesthetics from ggplot()?}

\item{show.legend}{Show legend for this layer?}
}
\value{
A ggplot2 layer
}
\description{
Add XRF Peaks to a ggplot
}
\keyword{datasets}
