% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_ray_binding_energies}
\alias{x_ray_binding_energies}
\alias{x_ray_cross_sections}
\alias{x_ray_emission_probabilities}
\title{EADL97 Atomic data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1612 rows and 5 columns.}
\source{
The PyMca5 python package (\url{https://pypi.org/project/PyMca5/#files}),
which obtained the data from the EADL page (\url{http://www-nds.iaea.org/epdl97/libsall.htm})
}
\usage{
x_ray_binding_energies

x_ray_cross_sections

x_ray_emission_probabilities
}
\description{
Data
}
\references{
D.E. Cullen, et al., "Tables and Graphs of Atomic Subshell and
Relaxation Data Derived from the LLNL Evaluated Atomic Data Library
(EADL), Z = 1 - 100," Lawrence Livermore National Laboratory, UCRL-50400,
Vol. 30, October 1991
}
\keyword{datasets}
