% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_ray_compton_profiles}
\alias{x_ray_compton_profiles}
\title{Compton profiles of the elements}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3162 rows and 4 columns.
}
\source{
DABAX copy vendored via the xraylib project (BSD-3-Clause,
  \url{https://github.com/tschoonj/xraylib}).
}
\usage{
x_ray_compton_profiles
}
\description{
The total electron momentum density \eqn{J(p_z)} per element (Biggs, Mendelsohn & Mann 1975,
Hartree-Fock; \eqn{p_z} in atomic units, \eqn{J} symmetric with \eqn{\int J\,dp_z = Z}). In the
impulse approximation the Compton-scattered energy distribution of a line \eqn{E_0} at angle
\eqn{\theta} is \eqn{\rho(E_2) \propto J(|p_z(E_2)|)\,|dp_z/dE_2|} with the Ribberfors relation
\deqn{p_z = 137.036\,\frac{E_0 - E_2 - E_0 E_2 (1-\cos\theta)/m c^2}
  {\sqrt{E_0^2 + E_2^2 - 2 E_0 E_2 \cos\theta}},}
which is the physical, asymmetric Doppler broadening of the Compton peak -- used by
\link{xrf_scatter_peaks} (\code{compton_shape = "profile"}) in place of a broadened Gaussian.
}
\references{
Biggs, F., Mendelsohn, L.B., Mann, J.B. (1975). Hartree-Fock Compton profiles for the elements.
\emph{Atomic Data and Nuclear Data Tables} 16, 201-309. \doi{10.1016/0092-640X(75)90030-3}

Ribberfors, R. (1975). Relationship of the relativistic Compton cross section to the momentum
distribution of bound electron states. \emph{Physical Review B} 12, 2067-2074.
\doi{10.1103/PhysRevB.12.2067}
}
\keyword{datasets}
