% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_ray_transitions}
\alias{x_ray_transitions}
\title{X-Ray transitions}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 4 columns.}
\usage{
x_ray_transitions
}
\description{
A data frame keeping track of the formal and siegbahn names of various transitions
}
\keyword{datasets}
