% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_ray_xrf_energies}
\alias{x_ray_xrf_energies}
\title{XRF quantification energies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1723 rows and 7 columns.
}
\source{
NIST (2018), EADL97. See \link{x_ray_energies} and \link{x_ray_emission_probabilities}.
}
\usage{
x_ray_xrf_energies
}
\description{
A form of \link{x_ray_energies} in a more suitable form for XRF quantification.
Best accessed via \link{xrf_energies}.
}
\references{
National Institute of Standards and Technology (NIST): X-Ray Transition Energies Database.
Retrieved August 2018. \url{https://physics.nist.gov/PhysRefData/XrayTrans/Html/search.html}.
}
\keyword{datasets}
