% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{xrf_add_baseline}
\alias{xrf_add_baseline}
\title{Add a baseline estimate from a function}
\usage{
xrf_add_baseline(.spectra, .fun, ..., .clamp = -Inf, .env = parent.frame())
}
\arguments{
\item{.spectra}{A spectra_df}

\item{.fun}{A function that receives a vector of values and outputs a vector of values
of the same length}

\item{...}{Passed to \code{.fun}, evaluated within .spectra}

\item{.clamp}{The minimum allowable value for background}

\item{.env}{The calling environment}
}
\value{
.spectra with a modified .spectra column
}
\description{
Add a baseline estimate from a function
}
