% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{xrf_add_baseline_als}
\alias{xrf_add_baseline_als}
\title{Asymmetric least squares baseline calcluation}
\usage{
xrf_add_baseline_als(spectra, p = 0.003, lambda = 6, maxit = 20,
  clamp = 0, values = .data$cps, env = parent.frame())
}
\arguments{
\item{spectra}{A data frame with a nested .spectra column}

\item{p, lambda, maxit}{See \link[baseline]{baseline.als}}

\item{clamp}{The smallest value that the baseline should be. Use -Inf to suppress.}

\item{values}{The value column in each spectra object}

\item{env}{The calling environment}
}
\value{
spectra with original information plus basline added
}
\description{
Uses the \link[baseline]{baseline.als} function in the baseline package.
}
