% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{xrf_add_baseline_pkg}
\alias{xrf_add_baseline_pkg}
\title{Caclulate baselines using the baseline package}
\usage{
xrf_add_baseline_pkg(
  .spectra,
  .values = .data$.spectra$cps,
  ...,
  .clamp = -Inf,
  .env = parent.frame()
)
}
\arguments{
\item{.spectra}{A spectra_df}

\item{.values}{The value column in each spectra object}

\item{...}{Passed to \link[baseline]{baseline}. Evaluated rowwise using objects in \code{spectra}.}

\item{.clamp}{The minimum allowable value for background}

\item{.env}{The calling environment}
}
\description{
Wraps a call to the \link[baseline]{baseline}(). See documentation in that package for
references for baseline correction.
}
\references{
See \link[baseline]{baseline}
}
