% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{xrf_add_smooth}
\alias{xrf_add_smooth}
\title{Smooth spectra using a function}
\usage{
xrf_add_smooth(.spectra, .fun, ..., .iter = 0, .epsilon = NA,
  .env = parent.frame())
}
\arguments{
\item{.spectra}{A spectra_df}

\item{.fun}{A function that receives a vector of values and outputs a vector of values
of the same length}

\item{...}{Passed to \code{.fun}, evaluated within .spectra}

\item{.iter}{Maximum number of iterations to smooth (minus 1)}

\item{.epsilon}{Further constrain maximum iterations}

\item{.env}{The calling environment}
}
\value{
A modified .spectra
}
\description{
Smooth spectra using a function
}
