% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{xrf_add_smooth_filter}
\alias{xrf_add_smooth_filter}
\title{Add smooth using iterative filtering}
\usage{
xrf_add_smooth_filter(
  .spectra,
  .values = .data$.spectra$cps,
  filter = xrf_filter_gaussian(),
  .iter = 5,
  .epsilon = -1,
  .env = parent.frame()
)
}
\arguments{
\item{.spectra}{A spectra_df}

\item{.values}{Which values to smooth}

\item{filter}{The filter which should be applied}

\item{.iter}{Maximum number of iterations to smooth}

\item{.epsilon}{Further constrain maximum iterations}

\item{.env}{The calling environment}
}
\description{
Add smooth using iterative filtering
}
