% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{xrf_add_smooth_gaussian}
\alias{xrf_add_smooth_gaussian}
\title{Add smooth using iterative gaussian smoothing}
\usage{
xrf_add_smooth_gaussian(.spectra, .values = .data$.spectra$cps, window = 3L,
  alpha = 2.5, tails = TRUE, .iter = 10, .epsilon = NA,
  .env = parent.frame())
}
\arguments{
\item{.spectra}{A spectra_df}

\item{.values}{Which values to smooth}

\item{window, alpha, tails}{See \link[smoother]{smth.gaussian}. Increasing the window size
(e.g., 5L or 10L) will increase the amount of smoothing.}

\item{.iter}{Maximum number of iterations to smooth (minus 1)}

\item{.epsilon}{Further constrain maximum iterations}

\item{.env}{The calling environment}
}
\description{
Add smooth using iterative gaussian smoothing
}
