% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{xrf_combine_spectra}
\alias{xrf_combine_spectra}
\alias{rbind.spectra}
\title{Combine XRF spectra}
\usage{
xrf_combine_spectra(..., .id = NULL)

\method{rbind}{spectra}(...)
}
\arguments{
\item{...}{XRF spectra or lists of XRF spectra.}

\item{.id}{Column in ouput that will contain names or indicies of input}
}
\value{
A spectra.
}
\description{
Behaviour is identical to \link[dplyr]{bind_rows}.
}
