% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{xrf_get_energies}
\alias{xrf_get_energies}
\title{Get XRF energies for selected elements}
\usage{
xrf_get_energies(elements = "everything", beam_energy_kev = Inf, ...)
}
\arguments{
\item{elements}{Elements or element lists (passed to \link{xrf_element_list}).}

\item{beam_energy_kev}{Beam energy}

\item{...}{Used to further \link[dplyr]{filter} the result.}
}
\value{
A subset of \link{xrf_energies}
}
\description{
Get XRF energies for selected elements
}
\examples{
xrf_get_energies("major", 25)

}
