% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra.R
\name{xrf_spectra}
\alias{xrf_spectra}
\title{Create an XRF spectra}
\usage{
xrf_spectra(.spectra, ..., .position = NA_integer_, .path = NA_character_,
  .validate = TRUE)
}
\arguments{
\item{.spectra}{A data frame or list of data frames with columns (at least) "energy_kev" and "cps".}

\item{...}{Key/value pairs, passed to \link[tibble]{tibble}.}

\item{.position}{If there is more than one spectrum in .path, a number describing which spectrum within .path}

\item{.path}{The file from whence the spectrum came. Can be NA.}

\item{.validate}{Use FALSE to skip validation}
}
\value{
A spectra df.
}
\description{
Create an XRF spectra
}
