% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_ray_energies}
\alias{x_ray_energies}
\title{Elemental X-ray energies}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1843 rows and 9 columns.}
\source{
Kaye, G. W. C., and T. H. Laby. Tables of Physical and Chemical Constants and Some Mathematical Functions.
16th edition. Essex, England; New York: Longman Sc & Tech, 1995. \url{http://www.kayelaby.npl.co.uk/toc/}.
}
\usage{
x_ray_energies
}
\description{
Emission and edge (minimum excitation) energies for elements, based on table 4.2.1 in Kaye and Laby (1995).
Additional references provided in the book are included in the references section.
}
\references{
Bearden (1967) Rev. Mod. Phys., 39(1), 78.

Clark (1963) Encyclopaedia of X-rays and Gamma Rays, Reinhold.

Dewey, Mapes and Reynolds (1969) Progress in Nuclear Energy, Series IX,
Analytical Chemistry, Elion and Stewart (eds), Pergamon.

Krause (1979) J. Physical and Chemical Reference Data, 8, 307–27.

Mitchell and Barfoot (1981) Nuclear Science Applications, 1, 99.

Robinson (ed.) (1974) Handbook of Spectroscopy, Vol. 1, p. 228, CRC Press, Cleveland.

Salem, S.I., S.L. Panossian, and R.A. Krause. "Experimental K and L Relative X-Ray Emission Rates."
Atomic Data and Nuclear Data Tables 14, no. 2 (August 1974): 91–109.
\url{https://doi.org/10.1016/S0092-640X(74)80017-3}.

Sandström (1957) Handbuch der Physik, Vol. XXX, X-rays, Springer-Verlag.

Singh et al. (1990) Nuclear Instruments and Methods in Physics Research, B51, 5.
}
\keyword{datasets}
