% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{xrf_absorption_jump}
\alias{xrf_absorption_jump}
\alias{xrf_absorption_jump_ratio}
\alias{xrf_transition_probability}
\alias{xrf_fluorescence_yield}
\alias{xrf_coster_kronig_probability}
\alias{xrf_relative_peak_intensity}
\title{Physical parameters of electron shell transitions}
\usage{
xrf_absorption_jump(element, shell, method = "viegele")

xrf_absorption_jump_ratio(element, shell, method = "viegele")

xrf_transition_probability(element, trans, method = "EADL97")

xrf_fluorescence_yield(element, shell, method = "EADL97")

xrf_coster_kronig_probability(
  element,
  shell,
  coster_kronig_trans,
  method = "EADL97"
)

xrf_relative_peak_intensity(element, shell, trans, beam_energy_kev = 50)
}
\arguments{
\item{element}{An element symbol or atomic number}

\item{shell}{A valid electron shell (K, L1, L2, L3). Invalid shells are silently ignored.}

\item{method}{Only the Viegele (1973) method is currently supported for xrf_absorption_jump,
and EADL97 tables are used for other parameters.}

\item{trans}{A transition (e.g., KL2).}

\item{coster_kronig_trans}{A sub-transition according to Coster-Kronig (1935).}

\item{beam_energy_kev}{The beam energy used to excite the electrons. Currently unused, but
should be incorporated into the calculation.}
}
\value{
A vector of numerics
}
\description{
See Details
}
\details{
Absoroption jump ratios: vectorized over both element and shell.
The absorption jump (often written \code{r} sub [K, L1, L2, L3])
can be used to calculate the probability that an electron from \code{shell} will be
ejected (J sub [K, L1, L2, L3]; the absorption jump ratio).

Transition probabilities are the probability that a specified transition takes
place rather than another in that shell. Generally given the letter "g" sub [KL2, etc.].
See also \link{x_ray_emission_probabilities}.

Fluorescence yields are the probability of emission of x-radiation rather than
an Auger electron. Generally given the letter omega sub [K, L1, L2, L3]. For
L shells, there are also Coster-Kronig (1935) transitions.

xrf_relative_peak_intensity attempts to combine these functions to estimate
the relative intensities of various peaks. This is an important input for
deconvolution, and needs to be tweaked based on sample size.
}
\examples{
# absorption jump ratio
xrf_absorption_jump("Pb", c("K", "L1", "L2", "L3"))

# probability that a K shell electron will be ejected
rK <- xrf_absorption_jump("Pb", c("K", "L1", "L2", "L3"))
JK <- (rK - 1) / rK
xrf_absorption_jump_ratio("Pb", c("K", "L1", "L2", "L3"))

}
\references{
D.E. Cullen, et al., "Tables and Graphs of Atomic Subshell and
Relaxation Data Derived from the LLNL Evaluated Atomic Data Library
(EADL), Z = 1 - 100," Lawrence Livermore National Laboratory, UCRL-50400,
Vol. 30, October 1991. \url{http://www-nds.iaea.org/epdl97/libsall.htm}.

Coster, D. and Kronig De L. (1935). New type of auger effect and its influence on the x-ray spectrum.
Pysica 2:13-24. \url{https://doi.org/10.1016/S0031-8914(35)90060-X}.

Veigele, W.J. (1973) Atomic and Nuclear Data Tables 5:51-111. p 54 and 55.
\url{https://doi.org/10.1016/S0092-640X(73)80015-4}.
}
