% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{xrf_add_baseline_pkg}
\alias{xrf_add_baseline_pkg}
\title{Caclulate baselines using the baseline package}
\usage{
xrf_add_baseline_pkg(spectra, ..., values = .data$cps, clamp = 0,
  env = parent.frame())
}
\arguments{
\item{spectra}{A data frame with a nested .spectra column}

\item{...}{Passed to \link[baseline]{baseline}. Evaluated rowwise using objects in \code{spectra}.}

\item{values}{The value column in each spectra object}

\item{clamp}{The smallest value that the baseline should be. Use -Inf to suppress.}

\item{env}{The calling environment}
}
\description{
Wraps a call to the \link[baseline]{baseline}()
}
