% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{xrf_add_baseline_snip}
\alias{xrf_add_baseline_snip}
\title{Calculate baselines using SNIP}
\usage{
xrf_add_baseline_snip(
  .spectra,
  .values = .data$.spectra$cps,
  ...,
  .clamp = -Inf,
  .env = parent.frame()
)
}
\arguments{
\item{.spectra}{A spectra_df}

\item{.values}{The value column in each spectra object}

\item{...}{Passed to \link[baseline]{baseline}. Evaluated rowwise using objects in \code{spectra}.}

\item{.clamp}{The minimum allowable value for background}

\item{.env}{The calling environment}
}
\description{
Based on \link[Peaks]{SpectrumBackground} (Peaks package). The most useful parameter to change is the
\code{iterations} argument (higher value will result in a more conservative baseline estimation).
}
\details{
... Passed to \link[Peaks]{SpectrumBackground}
}
