% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{xrf_add_deconvolution_gls}
\alias{xrf_add_deconvolution_gls}
\alias{xrf_deconvolute_gaussian_least_squares}
\title{Deconvolute a spectrum using gaussian least squares}
\usage{
xrf_add_deconvolution_gls(.spectra, .energy_kev = .data$.spectra$energy_kev,
  .values = .data$.spectra$cps - .data$.spectra$baseline,
  energy_min_kev = -Inf, energy_max_kev = Inf, peaks = xrf_energies(),
  sigma = 0.07, .env = parent.frame())

xrf_deconvolute_gaussian_least_squares(energy_kev, response,
  peaks = xrf_energies(), sigma = 0.07, energy_min_kev = -Inf,
  energy_max_kev = Inf)
}
\arguments{
\item{.spectra}{A \link{xrf_spectra} tibble.}

\item{.values, response}{Values to use for the deconvolution}

\item{energy_min_kev, energy_max_kev}{Constrain the energies that should be deconvoluted}

\item{peaks}{A data frame of columns with columns "element", "}

\item{sigma}{The standard deviation of the peaks.}

\item{.env}{Calling environment}

\item{energy_kev, .energy_kev}{Energy values corresponding to response/.values}
}
\value{
A modified version of .spectra
}
\description{
Deconvolute a spectrum using gaussian least squares
}
