% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xrf_energies}
\alias{xrf_energies}
\title{XRF quantification energies}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 287 rows and 8 columns.}
\source{
NIST (2018), EADL97. See \link{x_ray_energies} and \link{x_ray_emission_probabilities}.
}
\usage{
xrf_energies
}
\description{
Common energies used to quantify elements in XRF, including edge energies (energy
needed to produce that line on the spectrum).
}
\references{
National Institute of Standards and Technology (NIST): X-Ray Transition Energies Database.
Retrieved August 2018. \url{https://physics.nist.gov/PhysRefData/XrayTrans/Html/search.html}.
}
\keyword{datasets}
