% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/physics.R
\name{xrf_energies}
\alias{xrf_energies}
\title{Get XRF energies for selected elements}
\usage{
xrf_energies(
  elements = "everything",
  beam_energy_kev = 50,
  ...,
  min_relative_intensity = 0.01
)
}
\arguments{
\item{elements}{Elements or element lists (passed to \link{xrf_element_list}).}

\item{beam_energy_kev}{Beam energy, used to filter the list and calculate relative peak intensities.}

\item{...}{Used to further \link[dplyr]{filter} the result.}

\item{min_relative_intensity}{Smallest peak to include, relatie to the maximum peak for a given
element.}
}
\value{
A subset of \link{x_ray_xrf_energies} with some additional information based on the beam energy.
}
\description{
Get XRF energies for selected elements
}
\examples{
xrf_energies("major", 25)

}
